+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `equality` operator overload
 - `gc_clamp` property
+- `single_runs` property
 ## [0.1] - 2024-11-27
 ### Added
 - `MeltingTemperature` enum