+++ CHANGELOG.md
@@ -6,11 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- `equality` operator overload
+- `__eq__` overload
+- `__str__` overload
 - `gc_clamp` property
 - `single_runs` property
 ### Changed
-- `property` deleter & setter removed
+### Removed
+- `property` deleter & setter
 ## [0.1] - 2024-11-27
 ### Added
 - `MeltingTemperature` enum