+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `gc_clamp` property
 - `single_runs` property
 ### Changed
+- Test system modified
 ### Removed
 - `property` deleter & setter
 ## [0.1] - 2024-11-27