+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.2] - 2025-01-09
 ### Added
 - `__eq__` overload
 - `__str__` overload
@@ -31,5 +32,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `complement` method
 - `reverse` method
 
-[Unreleased]: https://github.com/openscilab/opr/compare/v0.1...dev
+[Unreleased]: https://github.com/openscilab/opr/compare/v0.2...dev
+[0.2]: https://github.com/openscilab/opr/compare/v0.1...v0.2
 [0.1]: https://github.com/openscilab/opr/compare/0baa8dd...v0.1